+++ watertap/costing/watertap_costing_package.py
@@ -11,8 +11,6 @@
 #
 ###############################################################################
 
-import functools
-
 import pyomo.environ as pyo
 
 from pyomo.util.calc_var_value import calculate_variable_from_constraint
@@ -45,10 +43,16 @@ from watertap.unit_models import (
     GAC,
 )
 
+from .util import (
+    make_capital_cost_var,
+    make_fixed_operating_cost_var,
+    cost_membrane,
+    cost_electrodialysis_stack,
+    cost_by_flow_volume,
+    cost_uv_aop_bundle,
+)
 
-class ROType(StrEnum):
-    standard = "standard"
-    high_pressure = "high_pressure"
+from .units.reverse_osmosis import cost_reverse_osmosis, ROType
 
 
 class PumpType(StrEnum):
@@ -70,65 +74,6 @@ class CrystallizerCostType(StrEnum):
     default = "default"
     mass_basis = "mass_basis"
     volume_basis = "volume_basis"
-
-
-def build_reverse_osmosis_cost_param_block(blk):
-
-    print("IN build_reverse_osmosis_cost_param_block")
-
-    costing = blk.parent_block()
-
-    blk.factor_membrane_replacement = pyo.Var(
-        initialize=0.2,
-        doc="Membrane replacement factor [fraction of membrane replaced/year]",
-        units=pyo.units.year**-1,
-    )
-    blk.membrane_cost = pyo.Var(
-        initialize=30,
-        doc="Membrane cost",
-        units=costing.base_currency / (pyo.units.meter**2),
-    )
-    blk.high_pressure_membrane_cost = pyo.Var(
-        initialize=75,
-        doc="Membrane cost",
-        units=costing.base_currency / (pyo.units.meter**2),
-    )
-
-
-# @staticmethod
-# TODO: move back into WaterTAPCosting class as a static method
-def register_costing_parameter_block(build_rule, parameter_block_name):
-    def register_costing_parameter_block_decorator(func):
-        @functools.wraps(func)
-        def add_costing_parameter_block(blk, *args, **kwargs):
-            parameter_block = blk.costing_package.component(parameter_block_name)
-            if parameter_block is None:
-                parameter_block = pyo.Block(rule=build_rule)
-                blk.costing_package.add_component(parameter_block_name, parameter_block)
-                # fix the parameters in case the build_rule did not
-                parameter_block.fix_all_vars()
-            elif parameter_block._rule is None or not hasattr(
-                parameter_block._rule, "_fcn"
-            ):
-                raise RuntimeError(
-                    "Use the register_costing_parameter_block decorator for specifying"
-                    "costing-package-level parameters"
-                )
-            elif parameter_block._rule._fcn is not build_rule:
-                other_rule = parameter_block._rule._fcn
-                raise RuntimeError(
-                    "Attempting to add identically named costing parameter blocks with "
-                    "different build rules to the costing package "
-                    f"{blk.costing_package}. Parameter block named "
-                    f"{parameter_block_name} was previously built by function "
-                    f"{other_rule.__name__} from module {other_rule.__module__}"
-                )
-            # else parameter_block was constructed by build_rule previously
-            return func(blk, *args, **kwargs)
-
-        return add_costing_parameter_block
-
-    return register_costing_parameter_block_decorator
 
 
 @declare_process_block_class("WaterTAPCosting")
@@ -951,38 +896,6 @@ class WaterTAPCostingData(FlowsheetCostingBlockData):
         )
 
     @staticmethod
-    @register_costing_parameter_block(
-        build_rule=build_reverse_osmosis_cost_param_block,
-        parameter_block_name="reverse_osmosis",
-    )
-    def cost_reverse_osmosis(blk, ro_type=ROType.standard):
-        """
-        Reverse osmosis costing method
-
-        TODO: describe equations
-
-        Args:
-            ro_type: ROType Enum indicating reverse osmosis type,
-                default = ROType.standard
-        """
-        if ro_type == ROType.standard:
-            membrane_cost = blk.costing_package.reverse_osmosis.membrane_cost
-        elif ro_type == ROType.high_pressure:
-            membrane_cost = (
-                blk.costing_package.reverse_osmosis.high_pressure_membrane_cost
-            )
-        else:
-            raise ConfigurationError(
-                f"{blk.unit_model.name} received invalid argument for ro_type:"
-                f" {ro_type}. Argument must be a member of the ROType Enum."
-            )
-        cost_membrane(
-            blk,
-            membrane_cost,
-            blk.costing_package.reverse_osmosis.factor_membrane_replacement,
-        )
-
-    @staticmethod
     def cost_pump(blk, pump_type=PumpType.high_pressure, cost_electricity_flow=True):
         """
         Pump costing method
@@ -1765,8 +1678,8 @@ WaterTAPCostingData.unit_mapping = {
     Pump: WaterTAPCostingData.cost_pump,
     EnergyRecoveryDevice: WaterTAPCostingData.cost_energy_recovery_device,
     PressureExchanger: WaterTAPCostingData.cost_pressure_exchanger,
-    ReverseOsmosis0D: WaterTAPCostingData.cost_reverse_osmosis,
-    ReverseOsmosis1D: WaterTAPCostingData.cost_reverse_osmosis,
+    ReverseOsmosis0D: cost_reverse_osmosis,
+    ReverseOsmosis1D: cost_reverse_osmosis,
     NanoFiltration0D: WaterTAPCostingData.cost_nanofiltration,
     NanofiltrationZO: WaterTAPCostingData.cost_nanofiltration,
     Crystallization: WaterTAPCostingData.cost_crystallizer,
@@ -1776,154 +1689,3 @@ WaterTAPCostingData.unit_mapping = {
     IonExchange0D: WaterTAPCostingData.cost_ion_exchange,
     GAC: WaterTAPCostingData.cost_gac,
 }
-
-
-def make_capital_cost_var(blk):
-    blk.capital_cost = pyo.Var(
-        initialize=1e5,
-        domain=pyo.NonNegativeReals,
-        units=blk.costing_package.base_currency,
-        doc="Unit capital cost",
-    )
-
-
-def make_fixed_operating_cost_var(blk):
-    blk.fixed_operating_cost = pyo.Var(
-        initialize=1e5,
-        domain=pyo.NonNegativeReals,
-        units=blk.costing_package.base_currency / blk.costing_package.base_period,
-        doc="Unit fixed operating cost",
-    )
-
-
-def cost_membrane(blk, membrane_cost, factor_membrane_replacement):
-    """
-    Generic function for costing a membrane. Assumes the unit_model
-    has an `area` variable or parameter.
-
-    Args:
-        membrane_cost - The cost of the membrane in currency per area
-        factor_membrane_replacement - Membrane replacement factor
-                                      [fraction of membrane replaced/year]
-    """
-
-    make_capital_cost_var(blk)
-    make_fixed_operating_cost_var(blk)
-    blk.membrane_cost = pyo.Expression(expr=membrane_cost)
-    blk.factor_membrane_replacement = pyo.Expression(expr=factor_membrane_replacement)
-
-    blk.capital_cost_constraint = pyo.Constraint(
-        expr=blk.capital_cost == blk.membrane_cost * blk.unit_model.area
-    )
-    blk.fixed_operating_cost_constraint = pyo.Constraint(
-        expr=blk.fixed_operating_cost
-        == blk.factor_membrane_replacement * blk.membrane_cost * blk.unit_model.area
-    )
-
-
-def cost_electrodialysis_stack(
-    blk,
-    membrane_cost,
-    spacer_cost,
-    membrane_replacement_factor,
-    electrode_cost,
-    electrode_replacement_factor,
-):
-    """
-    Generic function for costing the stack in an electrodialysis unit.
-    Assumes the unit_model has a `cell_pair_num`, `cell_width`, and `cell_length`
-    set of variables used to size the total membrane area.
-
-    Args:
-        membrane_cost - The total cost of the CEM and AEM per cell pair in currency per area
-
-        spacer_cost - The total cost of the spacers per cell pair in currency per area
-
-        membrane_replacement_factor - Replacement factor for membranes and spacers
-                                      [fraction of membranes/spacers replaced/year]
-
-        electrode_cost - The total cost of electrodes in a given stack in currency per area
-
-        electrode_replacement_factor - Replacement factor for electrodes
-                                        [fraction of electrodes replaced/year]
-    """
-    make_capital_cost_var(blk)
-    make_fixed_operating_cost_var(blk)
-
-    blk.membrane_cost = pyo.Expression(expr=membrane_cost)
-    blk.membrane_replacement_factor = pyo.Expression(expr=membrane_replacement_factor)
-    blk.spacer_cost = pyo.Expression(expr=spacer_cost)
-    blk.electrode_cost = pyo.Expression(expr=electrode_cost)
-    blk.electrode_replacement_factor = pyo.Expression(expr=electrode_replacement_factor)
-
-    blk.capital_cost_constraint = pyo.Constraint(
-        expr=blk.capital_cost
-        == (blk.membrane_cost + blk.spacer_cost)
-        * (
-            blk.unit_model.cell_pair_num
-            * blk.unit_model.cell_width
-            * blk.unit_model.cell_length
-        )
-        + blk.electrode_cost * (blk.unit_model.cell_width * blk.unit_model.cell_length)
-    )
-    blk.fixed_operating_cost_constraint = pyo.Constraint(
-        expr=blk.fixed_operating_cost
-        == blk.membrane_replacement_factor
-        * (blk.membrane_cost + blk.spacer_cost)
-        * (
-            blk.unit_model.cell_pair_num
-            * blk.unit_model.cell_width
-            * blk.unit_model.cell_length
-        )
-        + blk.electrode_replacement_factor
-        * blk.electrode_cost
-        * (blk.unit_model.cell_width * blk.unit_model.cell_length)
-    )
-
-
-def cost_by_flow_volume(blk, flow_cost, flow_to_cost):
-    """
-    Generic function for costing by flow volume.
-
-    Args:
-        flow_cost - The cost of the device in [currency]/([volume]/[time])
-        flow_to_cost - The flow costed in [volume]/[time]
-    """
-    make_capital_cost_var(blk)
-    blk.flow_cost = pyo.Expression(expr=flow_cost)
-    blk.capital_cost_constraint = pyo.Constraint(
-        expr=blk.capital_cost == blk.flow_cost * flow_to_cost
-    )
-
-
-def cost_uv_aop_bundle(blk, reactor_cost, lamp_cost, factor_lamp_replacement):
-    """
-    Generic function for costing a UV system.
-
-    Args:
-        reactor_cost - The cost of UV reactor in [currency]/[volume]
-        lamp_cost - The costs of the lamps, sleeves, ballasts and sensors in [currency]/[kW]
-    """
-    make_capital_cost_var(blk)
-    make_fixed_operating_cost_var(blk)
-    blk.reactor_cost = pyo.Expression(expr=reactor_cost)
-    blk.lamp_cost = pyo.Expression(expr=lamp_cost)
-    blk.factor_lamp_replacement = pyo.Expression(expr=factor_lamp_replacement)
-
-    flow_in = pyo.units.convert(
-        blk.unit_model.control_volume.properties_in[0].flow_vol,
-        to_units=pyo.units.m**3 / pyo.units.hr,
-    )
-
-    electricity_demand = pyo.units.convert(
-        blk.unit_model.electricity_demand[0], to_units=pyo.units.kW
-    )
-
-    blk.capital_cost_constraint = pyo.Constraint(
-        expr=blk.capital_cost
-        == blk.reactor_cost * flow_in + blk.lamp_cost * electricity_demand
-    )
-    blk.fixed_operating_cost_constraint = pyo.Constraint(
-        expr=blk.fixed_operating_cost
-        == blk.factor_lamp_replacement * blk.lamp_cost * electricity_demand
-    )
